Create a new component to the Head
